refactor: UserIcon 컴포넌트 수정
재사용을 위한 props 명 수정 및 common 폴더로 위치 이동

Related to #109
