+++ fe/src/components/issueList/issueTable/list/listItems/listItemRight/ListItemRight.jsx
@@ -1,16 +1,15 @@
 import React, { useContext } from 'react';
 import styled from 'styled-components';
-import UserIcon from './UserIcon';
-import { MainPageInfoContext } from '@pages/IssueList';
+import UserIcon from '../../../../../common/icon/UserIcon';
 
 const ListItemRightBox = styled.div`
   margin-right: 10px;
 `;
 
-const ListItemRight = ({ issue: { authorUrl } }) => {
+const ListItemRight = ({ issue: { author, authorUrl } }) => {
   return (
     <ListItemRightBox>
-      <UserIcon authorUrl={authorUrl} />
+      <UserIcon userImgUrl={authorUrl} alt={author} />
     </ListItemRightBox>
   );
 };
